refactor: Async 관련 메소드 다른 컴포넌트로 분리 #368
